backend/requirements.txt
fastapi==0.129.0
langchain_community==0.4.1
langchain_core==1.2.12
langchain_huggingface==1.2.0
langchain_openai==1.1.9
langchain_text_splitters==1.1.0
langgraph==1.0.8
pydantic==2.12.5
python-dotenv==1.2.1
tavily==1.1.0
uvicorn==0.40.0
chromadb
aiosqlite
